Quick Changes For Security & Convinience.
In Mod Commands I changed "accountban" to "ban"

On the Admin commands for security and convinience purposes I changed the following commands:

getid = id
pp = minime
senditem = give
item = get

This is how they have been renamed.
I also made a small change to the location of ::sz I know no one really uses this command but when I need to talk to players the current sz in live copy right now is glitched and you can not see what the other person is typing so I just changed the coords it sends you to.

In the Normalcommand.java I removed the lines for "item" and "pickup" removing the ability for normal admins to spawn items.

Welcome message, well just that.